+++ drivers/convex_decomp/SCsub
@@ -4,4 +4,13 @@ Import('env')
 
 env.add_source_files(env.drivers_sources, "*.cpp")
 
+# Thirdparty dependencies
+thirdparty_dir = "#thirdparty/b2d_convexdecomp/"
+thirdparty_sources = [
+	"b2Polygon.cpp",
+	"b2Triangle.cpp",
+]
+thirdparty_sources = [thirdparty_dir + file for file in thirdparty_sources]
+env.add_source_files(env.drivers_sources, thirdparty_sources)
+
 Export('env')
