+++ modules/openssl/stream_peer_openssl.h
@@ -30,16 +30,18 @@
 #ifndef STREAM_PEER_OPEN_SSL_H
 #define STREAM_PEER_OPEN_SSL_H
 
-#include "curl_hostcheck.h"
 #include "global_config.h"
 #include "io/stream_peer_ssl.h"
 #include "os/file_access.h"
-#include <stdio.h> // If you don't know what this is for stop reading now.
+
+#include "thirdparty/misc/curl_hostcheck.h"
 
 #include <openssl/bio.h> // BIO objects for I/O
 #include <openssl/err.h> // Error reporting
 #include <openssl/ssl.h> // SSL and SSL_CTX for SSL connections
 #include <openssl/x509v3.h>
+
+#include <stdio.h>
 
 class StreamPeerOpenSSL : public StreamPeerSSL {
 private:
