+++ scene/resources/SCsub
@@ -3,7 +3,6 @@
 Import('env')
 
 env.add_source_files(env.scene_sources, "*.cpp")
-env.add_source_files(env.scene_sources, "*.c")
 
 Export('env')
