+++ scene/resources/dynamic_font_stb.cpp
@@ -33,7 +33,6 @@
 
 #define STB_TRUETYPE_IMPLEMENTATION
 #include "os/file_access.h"
-#include "stb_truetype.h"
 
 void DynamicFontData::lock() {
